Use StyxServerProvider throughout. Move LoggingAssertion.kt in support package.
