+++ system-tests/ft-suite/src/test/kotlin/com/hotels/styx/routing/ConditionRoutingSpec.kt
@@ -15,13 +15,13 @@
  */
 package com.hotels.styx.routing
 
-import com.hotels.styx.StyxConfig
-import com.hotels.styx.StyxServer
 import com.hotels.styx.api.HttpHeaderNames.HOST
 import com.hotels.styx.api.HttpRequest.get
 import com.hotels.styx.api.HttpResponseStatus.OK
 import com.hotels.styx.client.StyxHttpClient
-import com.hotels.styx.startup.StyxServerComponents
+import com.hotels.styx.support.StyxServerProvider
+import com.hotels.styx.support.proxyHttpHostHeader
+import com.hotels.styx.support.proxyHttpsHostHeader
 import io.kotlintest.Spec
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.StringSpec
@@ -33,7 +33,7 @@ class ConditionRoutingSpec : StringSpec() {
     init {
         "Routes HTTP protocol" {
             val request = get("/11")
-                    .header(HOST, "${styxServer.proxyHttpAddress().hostName}:${styxServer.proxyHttpAddress().port}")
+                    .header(HOST, styxServer().proxyHttpHostHeader())
                     .build();
 
             client.send(request)
@@ -49,7 +49,7 @@ class ConditionRoutingSpec : StringSpec() {
 
         "Routes HTTPS protocol" {
             val request = get("/2")
-                    .header(HOST, "${styxServer.proxyHttpsAddress().hostName}:${styxServer.proxyHttpsAddress().port}")
+                    .header(HOST, styxServer().proxyHttpsHostHeader())
                     .build();
 
             client.secure()
@@ -63,7 +63,9 @@ class ConditionRoutingSpec : StringSpec() {
         }
     }
 
-    val yamlText = """
+    val client: StyxHttpClient = StyxHttpClient.Builder().build()
+
+    val styxServer = StyxServerProvider("""
         proxy:
           connectors:
             http:
@@ -99,19 +101,13 @@ class ConditionRoutingSpec : StringSpec() {
                   config:
                     status: 200
                     content: "Hello, from http server!"
-      """.trimIndent()
-
-    val client: StyxHttpClient = StyxHttpClient.Builder().build()
-
-    val styxServer = StyxServer(StyxServerComponents.Builder()
-            .styxConfig(StyxConfig.fromYaml(yamlText))
-            .build())
+      """.trimIndent())
 
     override fun beforeSpec(spec: Spec) {
-        styxServer.startAsync().awaitRunning()
+        styxServer.restart()
     }
 
     override fun afterSpec(spec: Spec) {
-        styxServer.stopAsync().awaitTerminated()
+        styxServer.stop()
     }
 }
