+++ system-tests/ft-suite/src/test/kotlin/com/hotels/styx/support/LoggingAssertion.kt
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.logging
+package com.hotels.styx.support
 
 import ch.qos.logback.classic.Level
 import ch.qos.logback.classic.spi.ILoggingEvent
